+++ packages/workspace-fusion/src/lib/integrations/garden/utils/configureHighlight.ts
@@ -1,3 +1,4 @@
+import { BaseEvent } from '@equinor/workspace-core';
 import { GardenController } from '@equinor/workspace-garden';
 import { FusionMediator } from '../../../types';
 
@@ -6,10 +7,11 @@ export function configureGardenHighlightSelection<
 	TData extends Record<PropertyKey, unknown>,
 	TExtendedFields extends string,
 	TCustomGroupByKeys extends Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown>
+	TContext extends Record<PropertyKey, unknown>,
+	TCustomSidesheetEvents extends BaseEvent = never
 >(
 	gardenController: GardenController<TData, TExtendedFields, TCustomGroupByKeys, TContext>,
-	mediator: FusionMediator<TData, TContext>
+	mediator: FusionMediator<TData, TContext, TCustomSidesheetEvents>
 ) {
 	mediator.selectionService.selectedNodes$.subscribe((val) => {
 		gardenController.selectedNodes.setValue(val.map(({ id }) => id));