+++ packages/workspace-fusion/src/lib/integrations/sidesheet/components/addSidesheet.tsx
@@ -1,19 +1,36 @@
+import { BaseEvent } from '@equinor/workspace-core';
 import { WorkspaceViewController } from '@equinor/workspace-react';
-import { FusionMediator, GetIdentifier, WorkspaceTabNames } from '../../../types';
+import { WorkspaceSidesheets, FusionMediator, WorkspaceTabNames } from '../../../types';
 import { SidesheetConfig } from '../sidesheet';
-import { SidesheetWrapper } from './wrapper';
+import { SidesheetAdvancedWrapper } from './wrapper';
+import { SidesheetSimpleWrapper } from './wrapper/SidesheetSimpleWrapper';
 
 export function addSidesheet<
 	TData extends Record<PropertyKey, unknown>,
 	TError,
-	TContext extends Record<PropertyKey, unknown> = never
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
 >(
-	config: SidesheetConfig<TData> | undefined,
+	config: SidesheetConfig<TData, TContext, TCustomSidesheetEvents> | undefined,
 	viewController: WorkspaceViewController<WorkspaceTabNames, TError>,
-	mediator: FusionMediator<TData, TContext>
+	mediator: FusionMediator<TData, TContext, TCustomSidesheetEvents>
 ) {
-	if (!config) return;
-	viewController.addSidesheetComponent(() => (
-		<SidesheetWrapper getIdentifier={mediator.getIdentifier} Component={config.Sidesheet} mediator={mediator} />
-	));
+	if (!config || Object.keys(config).length === 0) return;
+
+	mediator.selectionService.selectedNodes$.subscribe((val) => {
+		const node = val[0];
+		if (!node) return;
+
+		const ev: WorkspaceSidesheets<TData> = { type: 'details_sidesheet', props: { id: node.id, item: node.item } };
+		mediator.sidesheetService.sendEvent(ev);
+	});
+
+	if (config.type === 'advanced') {
+		viewController.addSidesheetComponent(() => <SidesheetAdvancedWrapper config={config} mediator={mediator} />);
+		return;
+	}
+	if (config.type === 'simple') {
+		viewController.addSidesheetComponent(() => <SidesheetSimpleWrapper config={config} mediator={mediator} />);
+		return;
+	}
 }