+++ packages/workspace-fusion/src/lib/integrations/sidesheet/types/workspaceProps.ts
@@ -1,5 +1,11 @@
+import { BaseEvent } from '@equinor/workspace-core';
+import { FusionMediator } from '../../../types';
 import { SidesheetConfig } from '../sidesheet';
 
-export type WorkspaceSidesheetProps<TData extends Record<PropertyKey, unknown>> = {
-	sidesheetOptions?: SidesheetConfig<TData>;
+export type WorkspaceSidesheetProps<
+	TData extends Record<PropertyKey, unknown>,
+	TContext extends Record<PropertyKey, unknown>,
+	TCustomSidesheetEvents extends BaseEvent
+> = {
+	sidesheetOptions?: SidesheetConfig<TData, TContext, TCustomSidesheetEvents>;
 };